+++ sktime/classification/feature_based/_signature_classifier.py
