--- sktime/classification/signature_based/__init__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Signature based time series classifiers."""
-
-__all__ = ["SignatureClassifier"]
-
-from sktime.classification.signature_based._signature_classifier import (
-    SignatureClassifier,
-)
